+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added support for Cardputer ADV
+- Added 802.15.4 packet capture (only on C5, C6)
 
 - **PN532 NFC Capability**
   - **NTAG Support (Type 2)**